+++ tests/sentry/integrations/github/test_integration.py
@@ -9,6 +9,7 @@ from django.urls import reverse
 import sentry
 from sentry.constants import ObjectStatus
 from sentry.integrations.github import API_ERRORS, GitHubIntegrationProvider
+from sentry.integrations.utils.repo import Repo, RepoTree
 from sentry.models import Integration, OrganizationIntegration, Project, Repository
 from sentry.plugins.base import plugins
 from sentry.plugins.bases import IssueTrackingPlugin2
@@ -147,6 +148,13 @@ class GitHubIntegrationTest(IntegrationTestCase):
             match=[responses.matchers.query_param_matcher({"per_page": pp, "page": 3})],
             json={"repositories": [repositories["baz"]]},
             headers={"link": ", ".join([gen_link(2, "prev"), first])},
+        )
+        # When the pagination code is not being tested
+        responses.add(
+            responses.GET,
+            url=self.base_url + "/installation/repositories",
+            match=[responses.matchers.query_param_matcher({"per_page": 100})],
+            json={"repositories": list(repositories.values())},
         )
 
         responses.add(
@@ -581,12 +589,12 @@ class GitHubIntegrationTest(IntegrationTestCase):
     def test_get_trees_for_org(self):
         """Fetch the tree representation of a repo"""
         expected_trees = {
-            "Test-Organization/bar": {"default_branch": "main", "files": []},
-            "Test-Organization/baz": {"default_branch": "master", "files": []},
-            "Test-Organization/foo": {
-                "default_branch": "master",
-                "files": ["src/sentry/api/endpoints/auth_login.py"],
-            },
+            "Test-Organization/bar": RepoTree(Repo("Test-Organization/bar", "main"), []),
+            "Test-Organization/baz": RepoTree(Repo("Test-Organization/baz", "master"), []),
+            "Test-Organization/foo": RepoTree(
+                Repo("Test-Organization/foo", "master"),
+                ["src/sentry/api/endpoints/auth_login.py"],
+            ),
         }
         with self.tasks():
             self.assert_setup_flow()
@@ -594,35 +602,34 @@ class GitHubIntegrationTest(IntegrationTestCase):
         integration = Integration.objects.get(provider=self.provider.key)
         installation = integration.get_installation(self.organization.id)
 
-        with patch.object(sentry.integrations.github.client.GitHubClientMixin, "page_size", 1):
-            assert not cache.get("githubtrees:repositories:Test-Organization")
-            # This allows checking for caching related output
-            self._caplog.set_level(logging.INFO, logger="sentry")
-            trees = installation.get_trees_for_org()
+        assert not cache.get("githubtrees:repositories:Test-Organization")
+        # This allows checking for caching related output
+        self._caplog.set_level(logging.INFO, logger="sentry")
+        trees = installation.get_trees_for_org()
 
-            # These checks are useful since they will be available in the GCP logs
-            expected_msg = "The Github App does not have access to Test-Organization/baz."
-            assert self._caplog.records[8].message == expected_msg
-            assert self._caplog.records[8].levelname == "ERROR"
-            # XXX: We would need to patch timezone to make sure the time is always the same
-            assert self._caplog.records[9].message.startswith("Caching trees for Test-Organization")
-            assert self._caplog.records[9].levelname == "INFO"
+        # These checks are useful since they will be available in the GCP logs
+        expected_msg = "The Github App does not have access to Test-Organization/baz."
+        assert self._caplog.records[6].message == expected_msg
+        assert self._caplog.records[6].levelname == "ERROR"
+        # XXX: We would need to patch timezone to make sure the time is always the same
+        assert self._caplog.records[7].message.startswith("Caching trees for Test-Organization")
+        assert self._caplog.records[7].levelname == "INFO"
 
-            assert cache.get("githubtrees:repositories:foo:Test-Organization") == [
-                {"full_name": "Test-Organization/foo", "default_branch": "master"},
-                {"full_name": "Test-Organization/bar", "default_branch": "main"},
-                {"full_name": "Test-Organization/baz", "default_branch": "master"},
-            ]
-            assert cache.get("githubtrees:repo:Test-Organization/foo") == {
-                "default_branch": "master",
-                "files": ["src/sentry/api/endpoints/auth_login.py"],
-            }
+        assert cache.get("githubtrees:repositories:foo:Test-Organization") == [
+            {"full_name": "Test-Organization/foo", "default_branch": "master"},
+            {"full_name": "Test-Organization/bar", "default_branch": "main"},
+            {"full_name": "Test-Organization/baz", "default_branch": "master"},
+        ]
+        assert cache.get("githubtrees:repo:Test-Organization/foo") == RepoTree(
+            Repo("Test-Organization/foo", "master"),
+            ["src/sentry/api/endpoints/auth_login.py"],
+        )
 
-            assert trees == expected_trees
+        assert trees == expected_trees
 
-            # Calling a second time should produce the same results
-            trees = installation.get_trees_for_org()
-            assert self._caplog.records[10].message == "Using cached trees for Test-Organization."
-            assert self._caplog.records[10].levelname == "INFO"
+        # Calling a second time should produce the same results
+        trees = installation.get_trees_for_org()
+        assert self._caplog.records[8].message == "Using cached trees for Test-Organization."
+        assert self._caplog.records[8].levelname == "INFO"
 
-            assert trees == expected_trees
+        assert trees == expected_trees
